move utils to @lib/utils folder
